pyproject.toml
[tool.poetry]
name = "fotogram"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
sqlalchemy = "^2.0.29"
fastapi = "^0.110.2"
uvicorn = "^0.29.0"
pydantic-settings = "^2.2.1"
bcrypt = "^4.1.3"
psycopg2-binary = "^2.9.9"
router = "^0.1"
passlib = "^1.7.4"
python-jose = "^3.3.0"
python-multipart = "^0.0.9"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.6.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
src = ["./src"]
select = []

fixable = ["ALL"]
unfixable = []
exclude = [".git", ".ruff_cache"]

line-length = 79
target-version = "py312"